requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
backports-zoneinfo==0.2.1
    # via
    #   -r requirements/base.txt
    #   django
boto3==1.34.109
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
botocore==1.34.109
    # via
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
click==8.1.7
    # via code-annotations
code-annotations==1.8.0
    # via -r requirements/test.in
coverage[toml]==7.5.1
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-appconf
    #   django-statici18n
    #   edx-i18n-tools
    #   openedx-django-pyfs
django-appconf==1.0.6
    # via
    #   -r requirements/base.txt
    #   django-statici18n
django-statici18n==2.5.0
    # via -r requirements/base.txt
dnspython==2.6.1
    # via
    #   -r requirements/base.txt
    #   pymongo
edx-i18n-tools==1.6.0
    # via -r requirements/base.txt
edx-opaque-keys==2.9.0
    # via -r requirements/base.txt
exceptiongroup==1.2.1
    # via pytest
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   -r requirements/base.txt
    #   openedx-django-pyfs
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via code-annotations
jmespath==1.0.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
lazy==1.6
    # via
    #   -r requirements/base.txt
    #   xblock
lxml[html-clean]==5.2.2
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   lxml-html-clean
    #   xblock
lxml-html-clean==0.1.1
    # via
    #   -r requirements/base.txt
    #   lxml
mako==1.3.5
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-utils
markupsafe==2.1.5
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
    #   xblock
openedx-django-pyfs==3.6.0
    # via
    #   -r requirements/base.txt
    #   xblock
packaging==24.0
    # via pytest
path==16.14.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pbr==6.0.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
polib==1.2.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pymongo==4.4.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pytest==8.2.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.8.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   botocore
    #   xblock
python-slugify==8.0.4
    # via code-annotations
pytz==2024.1
    # via
    #   -r requirements/base.txt
    #   xblock
pyyaml==6.0.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-i18n-tools
    #   xblock
s3transfer==0.10.1
    # via
    #   -r requirements/base.txt
    #   boto3
simplejson==3.19.2
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-utils
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   fs-s3fs
    #   python-dateutil
sqlparse==0.5.0
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.2.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.11.0
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   edx-opaque-keys
urllib3==1.26.18
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   botocore
web-fragments==2.2.0
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-utils
webob==1.8.7
    # via
    #   -r requirements/base.txt
    #   xblock
xblock[django]==4.0.1
    # via
    #   -r requirements/base.txt
    #   xblock-utils
xblock-utils==4.0.0
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
